Move database methods to models.
